requirements.txt
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.0
pyathena>=3.0.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.17.0
